moved functions around, new file too
