+++ src/memory.c
@@ -12,3 +12,12 @@ void	free_array(char **s)
 	}
 	free(s);
 }
+
+void	free_run(t_run *run)
+{
+	close_pipes(run->pipefds, run->len);
+	free_array(run->env->arr);
+	free(run->env);
+	free_list(run->cmd_list);
+	clear_history();
+}
